requirements.txt
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
dataclasses==0.8
decorator==4.4.2
dgl==0.5.3
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==4.0.1
install==1.3.4
joblib==1.0.0
mecab-python3==1.0.3
murmurhash==1.0.5
networkx==2.5
nltk==3.5
numpy==1.19.4
packaging==20.8
pandas==1.1.5
plac==1.1.3
preshed==3.0.5
pyparsing==2.4.7
pythainlp==2.3.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
tinydb==4.4.0
tokenizers==0.9.4
torch==1.7.0
tqdm==4.54.1
transformers==4.1.1
typing-extensions==3.10.0.0
unidic-lite==1.0.7
urllib3==1.26.2
wasabi==0.8.0
zipp==3.4.1
